fix: address review comments for record naming strategy PR
- Writer: use BeanUtils.fieldName() for records to avoid stripping is/get
  prefixes that are part of the field name in Java records
- Reader: apply namingStrategy to record parameterNames so deserialized
  JSON keys (e.g. currency_code) match the strategy-applied names
- Tests: move record tests to test-jdk17 module (Java 16+ required),
  replace local record and var with class-level record and explicit types,
  add round-trip deserialization assertions and boolean prefix test
