+++ core/src/test/java/com/alibaba/fastjson2/annotation/JSONTypeNamingSnake.java
@@ -1,8 +1,6 @@
 package com.alibaba.fastjson2.annotation;
 
 import com.alibaba.fastjson2.JSON;
-import com.alibaba.fastjson2.JSONFactory;
-import com.alibaba.fastjson2.PropertyNamingStrategy;
 import org.junit.jupiter.api.Tag;
 import org.junit.jupiter.api.Test;
 
@@ -55,29 +53,12 @@ public class JSONTypeNamingSnake {
     }
 
     @JSONType(naming = PropertyNamingStrategy.SnakeCase)
-    public record RecordDto(String currencyCode) {}
-
-    @JSONType(naming = PropertyNamingStrategy.SnakeCase)
     static class ClassDto {
         public String currencyCode = "840";
     }
 
     @Test
-    public void testRecord() {
-        assertEquals("{\"currency_code\":\"840\"}", JSON.toJSONString(new RecordDto("840")));
+    public void testClass() {
         assertEquals("{\"currency_code\":\"840\"}", JSON.toJSONString(new ClassDto()));
-    }
-
-    @Test
-    public void testRecordGlobalNaming() {
-        var provider = JSONFactory.getDefaultObjectWriterProvider();
-        var prev = provider.getNamingStrategy();
-        try {
-            provider.setNamingStrategy(PropertyNamingStrategy.SnakeCase);
-            record GlobalRecordDto(String currencyCode) {}
-            assertEquals("{\"currency_code\":\"840\"}", JSON.toJSONString(new GlobalRecordDto("840")));
-        } finally {
-            provider.setNamingStrategy(prev);
-        }
     }
 }
